// src/types/tableTypes.ts
export interface Survey {
    survey_id: number; // Уникальный идентификатор опроса
    region_id: number; // Регион, связанный с опросом
    survey_type: survey_type_enum; // Тип опроса
    topic: string; // Тематика опроса
    status: survey_status_enum; // Статус задания (забронировано или свободно)
    created_at: string; // Дата и время создания опроса (ISO строка)
    reserved_until: string; // Дата и время до какого зарегестр опрос
    updated_at: string; // Последнее обновление записи (ISO строка)
}
// ...
export type survey_type_enum  ='animals'| 'city'
export type survey_status_enum  ='reserved' | 'available' | 'in_progress'; // Статус задания